pyproject.toml
[project]
name = "miniondesk"
version = "2.0.0"
description = "MinionDesk — Model-agnostic enterprise AI assistant framework"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Host dependencies
    "python-telegram-bot>=21.0",
    "discord.py>=2.3.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    # Enterprise / LDAP
    "ldap3>=2.9",
]

[project.optional-dependencies]
claude = ["anthropic>=0.40.0"]
gemini = ["google-generativeai>=0.8.0"]
openai = ["openai>=1.55.0"]
all = [
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "openai>=1.55.0",
]

[project.scripts]
miniondesk = "run:main"

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["host*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top of file
    "F401",   # imported but unused (common in __init__.py)
    "F841",   # local variable assigned but never used
    "E711",   # comparison to None (use `is`)
    "E712",   # comparison to True/False
    "W291",   # trailing whitespace
    "W293",   # whitespace before ':'
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
